services/backend/requirements.txt
aerich==0.7.1
asyncpg==0.27.0
email-validator==2.1.0
fastapi==0.88.0
tortoise-orm==0.19.2
uvicorn==0.20.0